# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * A.N.O.T.H.E.R. Private API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  online: 'online',
  away: 'away',
  dnd: 'dnd',
  offline: 'offline',
} as const;

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface User {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  status: UserStatus;
  role: UserRole;
  createdAt: string;
}

export type UserProfileStatus = typeof UserProfileStatus[keyof typeof UserProfileStatus];


export const UserProfileStatus = {
  online: 'online',
  away: 'away',
  dnd: 'dnd',
  offline: 'offline',
} as const;

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface UserProfile {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  status: UserProfileStatus;
  role: UserProfileRole;
  createdAt: string;
}

export interface AdminUser {
  id: number;
  username: string;
  displayName: string;
  status: string;
  role: string;
  banned: boolean;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface Server {
  id: number;
  name: string;
  /** @nullable */
  iconUrl?: string | null;
  ownerId: number;
  memberCount: number;
  createdAt: string;
}

export type ServerMemberRole = typeof ServerMemberRole[keyof typeof ServerMemberRole];


export const ServerMemberRole = {
  owner: 'owner',
  admin: 'admin',
  member: 'member',
} as const;

export interface ServerMember {
  id: number;
  serverId: number;
  userId: number;
  role: ServerMemberRole;
  joinedAt: string;
  user: UserProfile;
}

export interface Channel {
  id: number;
  serverId: number;
  name: string;
  createdAt: string;
}

export interface Message {
  id: number;
  channelId: number;
  userId: number;
  content: string;
  /** @nullable */
  editedAt?: string | null;
  /** @nullable */
  deletedAt?: string | null;
  createdAt: string;
  author: UserProfile;
}

export interface InviteCode {
  code: string;
  generatedById: number;
  /** @nullable */
  usedById?: number | null;
  revoked: boolean;
  createdAt: string;
  /** @nullable */
  usedAt?: string | null;
}

export interface AdminStats {
  userCount: number;
  messageCount: number;
  serverCount: number;
  activeUserCount: number;
}

export interface AvatarUploadResult {
  url: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface RegisterInput {
  /**
     * @minLength 3
     * @maxLength 32
     */
  username: string;
  /** @minLength 8 */
  password: string;
  /**
     * @minLength 1
     * @maxLength 50
     */
  displayName: string;
  inviteCode?: string;
}

export interface ServerInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}

export interface ChannelInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  name: string;
}

export interface MessageInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
}

export interface MessageUpdate {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  content: string;
}

export type ProfileUpdateStatus = typeof ProfileUpdateStatus[keyof typeof ProfileUpdateStatus];


export const ProfileUpdateStatus = {
  online: 'online',
  away: 'away',
  dnd: 'dnd',
  offline: 'offline',
} as const;

export interface ProfileUpdate {
  /**
     * @minLength 1
     * @maxLength 50
     */
  displayName?: string;
  /** @maxLength 200 */
  bio?: string;
  status?: ProfileUpdateStatus;
}

export type MemberRoleUpdateRole = typeof MemberRoleUpdateRole[keyof typeof MemberRoleUpdateRole];


export const MemberRoleUpdateRole = {
  admin: 'admin',
  member: 'member',
} as const;

export interface MemberRoleUpdate {
  role: MemberRoleUpdateRole;
}

export interface InviteInput {
  note?: string;
}

export type ListMessagesParams = {
/**
 * Cursor — return messages before this message ID
 */
before?: number;
/**
 * Max messages to return (default 50)
 */
limit?: number;
};