# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: A.N.O.T.H.E.R. Private API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register with invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid data or invite code
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /admin/invites:
    get:
      operationId: listInvites
      tags: [admin]
      summary: List all invite codes
      responses:
        "200":
          description: Invite codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InviteCode"
    post:
      operationId: generateInvite
      tags: [admin]
      summary: Generate a new invite code
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "201":
          description: Invite code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteCode"
  /admin/invites/{code}:
    delete:
      operationId: revokeInvite
      tags: [admin]
      summary: Revoke an invite code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Invite revoked
...
  /admin/users/{userId}/kick:
    post:
      operationId: kickUser
      tags: [admin]
      summary: Kick a user from all servers
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User kicked
  /admin/users/{userId}/ban:
    post:
      operationId: banUser
      tags: [admin]
      summary: Ban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User banned
  /admin/users/{userId}/unban:
    post:
      operationId: unbanUser
      tags: [admin]
      summary: Unban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User unbanned
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, status, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, away, dnd, offline]
        role:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
          format: date-time
...
    InviteCode:
      type: object
      required: [code, generatedById, revoked, createdAt]
      properties:
        code:
          type: string
        generatedById:
          type: integer
        usedById:
          type: ["integer", "null"]
        revoked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        usedAt:
          type: ["string", "null"]
          format: date-time
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    RegisterInput:
      type: object
      required: [username, password, displayName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        password:
          type: string
          minLength: 8
        displayName:
          type: string
          minLength: 1
          maxLength: 50
        inviteCode:
          type: string
...
    InviteInput:
      type: object
      properties:
        note:
          type: string